Clean up and clarify Pulumi container commands in docker-compose.yml
There are a few things going on here, but it's all in service of being
more strict, clear, and extensible with what is going on in our Pulumi
container. These approaches should also be extended to our other
Docker Compose services, but this deals just with the Pulumi container.

First, we set an explicit `ENTRYPOINT` for our Pulumi container,
rather than passing a command starting with `/bin/bash -c`. This
protects us in case our base container ever changes and sets its own
`ENTRYPOINT` (our current one happens to not have one, though).

(With the changes to the command that we're making, we also _have_ to
set an entrypoint, because the long string we're generating is not
going to be an executable that exists in the container.)

We also get a bit more leeway to tighten up exactly how we invoke
`bash` here. We're adding the `-o errexit` and `-o nounset` options to
the entrypoint, meaning that any command failures (handy for multiple
commands!) will immediately fail the container, as will any unset
variables. I've previously seen failures of the Pulumi code that ended
up _not_ short-circuiting the whole `docker-compose` run for our
tests, making it difficult to track down the problem. Now, everything
fails fast.

(I could have just used `/bin/bash -euc`, but I generally favor being
more explicit, particularly in this scenario.)

Next, we express our command using the `- |` trick. We're using array
syntax, rather than a string (triggered by `-`), and we're preserving
any newlines in the following string (triggered by the `|`). By using
a single-element array, this entire string is passed to `bash -c` and
invoked. Expressing the command like this means we don't have to worry
about whether or not we need `&&` or `;` between commands, or anything
else like that; we can just write a straightforward bash script. We
also don't have to worry about extra levels of quoting, which has
often confused my editor's formatting.

Because of this, we also unlock the ability to do some
repetition-saving loops in our command. At the moment, this is
overkill, since we're only exporting a single stack output. However,
future PRs will be adding to this; by expressing this in a loop, we'll
make it easier to extend, and cut down on copy-paste errors.

Finally, the `SHELL` directive in `Dockerfile.pulumi` is moved closer to
where it's actually needed, for clarity.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>