Fix up FargateService abstraction w/r/t entrypoint and command
Previously, we were simply passing a single string as a command to
most of our Fargate containers.

However, with the recent change of our Rust containers to have an
entrypoint, rather than a command, these will no longer work; we would
be executing `/analyzer-dispatcher /analyzer-dispatcher`, for example.

We now remove our existing uses of `command` in our Fargate services,
because now they just aren't needed.

Though strictly not necessary at the moment, the type of `command` was
changed from `Optional[str]` to `Optional[List[str]]`, to better
reflect the underlying primitives. Additionally, a corresponding
`entrypoint` argument was also added, in case we should ever need to
make use of that in the future.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>