Pass in values for dgraph, swarm config buckets; dgraph logs group
In order to reduce shared knowledge of infrastructure naming patterns,
we begin to start passing in values to `graplctl` for the following:

- swarm-config-bucket
- dgraph-config-bucket
- dgraph-logs-group

Previously, all of these names were pieced together internally by
prepending a "deployment name" to a fixed suffix. This spreads
knowledge of names across both Pulumi and `graplctl`, making it more
difficult to change.

Now, `graplctl` simply operates directly on the information it has
been given, decoupling it from the rest of our code.

For each command that requires these pieces of data, a new CLI
argument has been added, along with an accompanying environment
variable that can be used instead.

Additionally, each of these values also has a corresponding Pulumi
stack output, making it easier to discover exactly what values should
be passed to `graplctl`.

It was an explicit design decision here to _not_ make `graplctl` aware
of how to manipulate Pulumi stacks directly. This keeps `graplctl` as
focused as possible, and does not introduce complex logic and
implementation details for functionality that is slated to be removed
soon anyway as we migrate this kind of infrastructure management code
to Nomad, etc.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>