Add graplctl-pulumi.sh wrapper script
Rather than having to manually specify generated infrastructure values
when invoking `graplctl`, and without having to build Pulumi-awareness
into `graplctl` itself, we instead introduce a simple wrapper script
that can inject the outputs of a Pulumi stack into the environment and
then call `graplctl`. The environment variables correspond to values
that `graplctl` recognizes, thus reducing the overhead for users for
running `graplctl`.

The script itself is fully documented, and is a drop-in replacement
for `graplctl`.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>