Stop explicitly managing network interfaces for our EC2 instances
The way we were managing things before this PR was causing problems
when the EC2 instances needed to be replaced (because we updated the
AMI they were using, for instance). The replacement instance would
come up, and then try to attach the specified network
interface. Unfortunately, this was still attached to the old instance,
which was still running.

One approach would be to use a [NetworkInterfaceAttachment][1] to
decouple the interface from the instance. However, this appears only
to be useful for managing secondary network interfaces, and not
primary ones.

That may be fine; however, we're opting to go for a simpler approach
here, and just configure the primary interface at the instance
level. This means that the interface just won't be shared at all.

I've confirmed that this code allows us to replace the EC2 instances
when we change the AMI.

[1]: https://www.pulumi.com/docs/reference/pkg/aws/ec2/networkinterfaceattachment/

Signed-off-by: Christopher Maier <chris@graplsecurity.com>