Create a ServiceConfiguration abstraction to interact with queues
This is a preliminary step for bundling up everything needed for a
"default" service or "retry" service to interact with our
`ServiceQueue`.

In particular, this clarifies and emphasizes the relationship a
"default" service has to the queues (pull from main queue, write to
retry queue), contrasted with that of a "retry" service (pull from the
retry queue, write to the dead-letter queue).

Signed-off-by: Christopher Maier <chris@graplsecurity.com>