Download all artifact files from the $ARTIFACT_FILE_DIRECTORY
Since we upload the files to the same directory, we can just pull them
all down with a glob. This makes the whole system more general and
robust; we no longer have to pass files explicitly to the
`merge_artifact_files.sh` script. We thus avoid the problems of
forgetting files, misspelling their names, etc.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>