Provide a TIMEOUT_SECS variable to adjust timeouts for e2e tests
Rather than providing a fixed timeout in our AWS e2e tests (to account
for the different times it takes to run there), we'll instead provide
a way to change the timeout values within the code based on an
environment variable.

This will be a better approach, because it takes some of the guesswork
out of the problem, and also allows the tests to take only as long as
they need.

It does allow our LensView query to run longer (300 vs 120 seconds) in
our Docker Compose environment, but I felt the simplicity of having
one timeout value outweighed the complexity of multiple timeouts, or
computing relative timeouts.

Apart from that, this should behave identically in our Docker Compose
setup.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>