Add some makefile targets to aid UX development against Local Grapl
There have been a lot of changes in how Grapl runs on a local
workstation. This commit introduces a few new makefile targets and
supporting code to help automate as much as possible to make it easier
to iteratively develop the Grapl UI locally.

Documentation is added in the new `LOCAL_UI_WORKFLOW.md` file.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>