Configure the Stale bot to close old PRs
This enables the Stale bot (https://github.com/probot/stale) to close
PRs that have not had activity in a while.

As configured, a PR must have one week (7 days) of inactivity before
being marked as "stale". If it experiences one more day without any
interaction, it will be closed.

This is in the interest of keeping our PR list up to date, and make it
easy to see what is currently being worked on and reviewed.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>