Abstract the calling of our various "check" services
The HCL-related services run as the local user (whether they need to
write to the filesystem or not).

The `buf` commands, on the other hand, do not; those containers seem
to require this, as they try to write to an container-internal
`/.cache` directory. At some point we can try to lock that down a bit
more; for now, though, this abstraction will help us see what we're
doing.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>