Remove execution_mode configuration from our PEXes
For an as-yet-undiagnosed reason, the changes in the PEX library that
were introduced to Pants in version 2.7.1rc0 made it so our PEXes that
were configured to run in "venv" mode no longer worked. They
apparently could not locate some of our Python modules, despite the
fact they were indeed packaged up properly.

The "zipapp" mode of execution (where the PEX is unzipped to disk and
then executed) appears to work, however; it's also the default
behavior for Pants-built PEX files.

So we'll just use that.

Background:
- https://pypi.org/project/pantsbuild.pants/2.7.1rc0/

Signed-off-by: Christopher Maier <chris@graplsecurity.com>