Add cargo audit configuration file
For the time being, we'll ignore the warnings and vulnerabilities that
are currently being flagged by `cargo-audit`. Many of them we can't
really do anything about anyway, given that there either is no fix, or
they are coming in from our dependencies.

The ones we do have control over have corresponding issues to ensure
we fix them.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>