Pull Pulumi artifacts from testing repository
First, we add a promotion step to our merge pipeline; if all our
containers build successfully, we'll use our new Cloudsmith Buildkite
plugin to promote the new artifacts out of the `raw` repository into
the `testing` one. Anything left in the `raw` repository will be
deleted over time, because anything that remains in there more than a
few minutes is by definition part of a failing commit.

Then, we parameterize our Pulumi stack with the value of this
repository, so we can instruct Nomad where to pull the containers
from. Previously, we specified this using a combination of "registry"
and "repo". However, because of how the naming in Cloudsmith works, we
always had to interject a `grapl/` into the resulting address. (That
is, `registry` was `docker.cloudsmith.io` and `repository` was `raw`,
with a resulting value of
`docker.cloudsmith.io/grapl/raw/<MY_IMAGE_NAME>:<TAG>`.

The fact that we specified a name of `grapl/<MY_IMAGE_NAME>` in our
Docker Compose build files complicated things. We did this to adhere
to Docker Hub conventions, even though we don't ever push these images
to Docker Hub. Therefore, we had to take special care to assemble the
correct name of an image, depending on whether we were deploying into
AWS or "local Grapl".

In an attempt to simplify things, we now just specify everything up to
the image name (so, e.g., `docker.cloudsmith.io/grapl/testing`) in a
Pulumi stack configuration variable, named
`container_repository`. Though it's technically a combination of a
registry and part of a repository, it does make things a bit easier to
reason about, and to configure.

Since we pin our artifacts in Pulumi using just the bare container
name and version, these will continue to work as before (because the
same names and versions will be present in the new container repository.

To make this work with local containers, we just remove the `grapl/`
prefix from the image name, making them the same as what we expect
elsewhere. This still works, and is actually a little bit safer
because now it will be impossible to accidentally push them to Docker
Hub. (I renamed our local `localstack` image to
`localstack-grapl-fork` to make it more obvious that we are using
something different than the "normal" localstack container.)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>