Remove old pinned artifacts
We don't currently have an automated way to *remove* unused artifacts
from our Pulumi stack. With all the internal revisions that have
occurred lately, we were left with quite a number of obsolete
packages.

Once this is merged into the `rc` branch, this will trigger a new run
of the `grapl/provision` pipeline.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>