Unset AWS credentials before trying to upload artifacts file
This is a wretchedly ugly hack but is currently the most
straightforward way to upload the resulting file to our artifacts
bucket.

The problem comes from our use of the assume-role plugin to inject AWS
credentials that have access to the staging account; this then gets in
the way of access to our artifacts S3 bucket, which resides in our CD
account.

(The irony is that we don't even need to interact with AWS for this at
all, but we need access to Pulumi in order to extract data from the
stack. I'll continue looking into better solutions, but for now, this
ought to work.)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>