Change ownership of PEX files in container images
In Pants 2.8, PEX files were created with all executable bits set; in
Pants 2.9, apparently only the file owner has execute
permissions. This broke our Python containers, because the `grapl`
user inside them could not execute any of the PEX files.

Now, we change ownership of the files as we copy them into the
container image.

We also change the `COPY` destination to `.`, since the `WORKDIR` has
already been set to `/home/grapl`, and the `CMD` implicitly references
that.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>