Switch from artifact-uploaders queue to docker for promotions
We recently retired the artifact-uploaders queue, since it was used
for uploading lambda zip files to Cloudsmith, and we no longer use
lambda functions.

However, they were also being used for promoting artifacts from one
repository to another. For that, we can just use the `docker` queue,
which also has access to the Cloudsmith API key, which is the main
thing we care about here.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>