Remove unnecessary plugin for deploy pipeline
The dump artifacts script is primarily used for getting all the Nomad
container logsprimarily for debugging failed tests. However, this is
just a pulumi deploy so there aren't multiple logs we should be
collecting, just one log that goes to standard out.

Additionally this has a fix for a copy paste error, where the stack
wasn't updated. Oh well, that's what code review is for