Refactor how we download CNI plugins
There are a few things addressed here:

First, rather than using `tar` to move things into their final
destinations, we'll just extract archives to a temporary directory
that doesn't require `root` permissions, and then do a `sudo mv`
afterwards. `tar` is far more complex than `mv`, and thus a liklier
source of potential vulnerabilities.

Second, as a result of the first change, we discovered that the
Firecracker driver-provided CNI plugins were not being installed
properly. Those binaries are provided in an `opt/cni/bin` directory
_inside_ the tarball, and we were not accounting for this when we were
installing them. As a result, they were ultimately being installed in
`/opt/cni/bin/opt/cni/bin`, and thus never actually used.

We also discovered that these were duplicating the upstream CNI
plugins we were also installing (those binaries are at the root of
their tarball, so they got properly installed in
`/opt/cni/bin`). Further digging revealed that the Firecracker
driver-provided plugins were several months behind the upstream
plugins. While they have a 3-line patch[^1] applied to them, we've
decided for now to stick with the upstream CNI plugins. Due to the
installation issues mentioned above, we were never actually using the
Firecracker driver-provided plugins. If we find in the future that we
need to switch, we can do so.

Finally, a bit more verbosity was added to the
`download_and_install_tarball` helper function, particularly the use
of `tree` to reveal the contents of the target installation directory
after all the binaries are in place. This increases visibility and
will help with debugging in the future (say, if/when we get any
tarballs with a more complex directory structure).

[^1]: https://github.com/cneira/plugins/commit/157b4539f75537bda000963a313601b7146107c4

Signed-off-by: Christopher Maier <chris@graplsecurity.com>