Only promote artifacts if they're not in the destination already
Due to what appears to be a Cloudsmith bug, we seem to be losing
images (tags, really) when we promote containers from our `raw`
repository to the `testing` repository. Cloudsmith detects that the
bytes of the image exist in the destination and thinks its job is
done; it doesn't seem to carry tags over, however. This manifests by
us not being able to find images in the `testing` repository, even
though it looks like Cloudsmith promoted them.

What we'll do now is add an additional check to help us assemble our
final list of artifacts to be promoted. We'll query the `testing`
repository to see if an image with the same sha256 sum as the one we
just built already exists there. If it does, then we must be dealing
with a service that hasn't changed since it was last pushed. In this
case, we'll just ignore it, and continue to use what's in the
`testing` repository. If no such image exists, then what we just built
must be new, so we will add it to our list of things to promote.

We have to add this logic here, rather than, say, in our Cloudsmith
plugin, because this list of services is also folded into our Pulumi
stack configuration files. Additionally, at some future time, we may
be able to simply inspect our code to see what changed, rather than
having to build images and push them first. All this argues for
putting the logic here, rather than any place else.

(We could use Pants to ask whether any of our Python images need to be
built, but the logic in this PR works regardless of what language the
services are written in. Were we to use Pants for this now, we would
also need to make changes to how we build the images; that's not
something we necessarily need to dig into right now.)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>