Remove yarn release file and yarnpath to it
Now that we're setting the yarn version explicitly in the Dockerfile
we don't need to keep a yarn binary around. This removes it and removes
the reference to it.
Keep in mind that this may break local development if anyone is doing
that since we're only specifying the version in the Dockerfile