Ensure failure of Prettier actually fails
Previously, only failures in the final TOML linting or formatting step
would cause the entire call to fail.

As coded, this will stop all operations at the first failure. We will
eventually want to either consolidate all exit codes, or break this up
such that only a single file format is operated on at a time.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>