Remove all fall-back values for TAG in Compose files
If every instance of TAG has its own fallback value, it can make it
difficult to know definitively where the value is coming from.

By not having a fallback, we require the value to be present when
`docker-compose` is run, providing a much stronger indication of where
the value ultimately comes from.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>