Disable the `codecov/patch` status check
This status check annotates lines of a PR that are introduced but not
covered by tests. Currently, this causes a lot noise, particularly for
Pulumi code.

I don't see a way to keep `codecov/patch` enabled, but for it to
ignore everything under `pulumi`; if we could sort that out, it may be
a preferable solution. For the time being, however, we'll just stop
doing the patch status checks.

It would be good to re-enable them in the future, however. Time for
more research!

Signed-off-by: Christopher Maier <chris@graplsecurity.com>