Rename `TAG` and `VERSION` to `IMAGE_TAG`
The `TAG` variable we used everywhere is now `IMAGE_TAG`, to be a bit
more descriptive and easier to `grep` for ("TAG" is really common,
particularly as a suffix).

The `VERSION` variable in `docker-bake.hcl` was also renamed to
`IMAGE_TAG` to align with this; its default value was also
removed. This means that the `IMAGE_TAG` variable set in the
`Makefile` is the One Source of Truth for this variable.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>