Disentangle and make explicit the dependencies for building images
Some of our existing build targets were obscuring the underlying
dependency structure somewhat, making it difficult to discern what was
going on. When it came to building images in CI to push to Cloudsmith,
we were actually doing more work than necessary, building `pulumi`,
`localstack` and `postgres` containers that we were never going to
push anywhere.

Here's an overview of what we have now:

- A new `build-ui` target, which builds only the `grapl-web-ui`
service. This is called out explicitly because this is the only thing
that depends on `build-engagement-view` (we build the website assets
and then copy them over to the Rust directory hierarchy, at which
point we bake them into the `grapl-web-ui` binary that is built.

- A new `build-grapl-services` target. This builds the rest of our
services that the SaaS will use. This depends on the aforementioned
`build-ui` target, so we still ensure that service is built. We also
pull in the PEX building command into this target, rather than using a
standalone target; if something else needs that logic, we can break it
out again, but this helps to contain some "target sprawl". The
explicit listing of the targets to build is unfortunate, but it has
the benefit of being explicit; an upcoming change to convert to
Buildkit's HCL file format will clean this up further.

- A new `build-local-services` target. Since we don't need to build
the local services in CI when we're pushing to Cloudsmith, it makes
sense to move these to their own target, so we can get better control
of what is built in different scenarios.

- A new `build-for-push` target. This captures the subset of things we
need to build and push to Cloudsmith. This is now used in CI.

- The `build-test-e2e` target now _only_ builds the E2E test
runner (previously it also caused everything else to be built,
obscuring the true dependency structure).

- The `test-e2e` and `test-integration` targets now explicitly depends
on the build steps they need.

- A new internal-use-only `_up` target has been extracted. `up`
depended on `build`, which was also a dependency of `test-integration`
and `test-e2e`. However, the latter two ended up invoking `make up`
via the `test-with-env` target. This means that `make build` was
effectively being invoked twice. This is no longer the case.

- The old `build` and `build-docker-images` targets have been removed
as obsolete.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>