some protobuf updates (#1466)
This PR updates the python-proto types to better reflect our intended data model. We now generically parametrize the SerDe base class with the protoc generated type, which allows each application type in python-proto to know how to serialize and deserialize itself. Additionally, we've added a SerDeWithInner base class to account for types (like Envelope) which contain an interior google.protobuf.Any. The upshot of all this is we can now parametrize the application type Envelope[T] where T is the type of the message contained in the Envelope's google.protobuf.Any wrapper.

Unfortunately, due to some tangly ergonomic issues in our rust codebase, the protocol buffer changes associated with this move can't be made just yet. As such, the extant Envelope has been left in place for now, and the changes described above have been made in a "v1beta2.NewEnvelope".

Subsequent changes will update the rust-proto crate to better reflect how the python-proto crate works, and eventually as we move services over to Kafka and get rid of the tangly serialization code we'll be able to remove the temporary "New" nomenclature and unify everything.