Use pulumi to create consul config files
By using pulumi instead of config files, we are able to quickly make
changes to consul instead of having to wait to restart all the consul
servers. This creates a default global config, and creates one
service-specific config to override the default. Unfortunately,
using nomad to specify the protocol doesn't work in this instance