Explicitly model dependencies for `build-test-unit`
Previously, `build-test-unit` did not formally depend on the `build-test-unit-js`
and `build-test-unit-rust` targets, even though it does depend on them
semantically (our pattern is that the "foo" target is a shorthand for
running all "foo-X" targets).

This is because we were allowing Docker to handle the parallelization
of the builds, rather than `make`.

This makes it difficult to ensure that our `make` targets are correct,
that none are left "dangling", and that we have no unexercised
targets. It also makes refactoring our Makefile targets tricky, as we
must rely on implicit knowledge of the dependency graph, rather than
what is explicitly written into the Makefile itself.

Furthermore, `make` can parallelize jobs with the `-j`/`--jobs`
option.

In the interest of explicitness, we now model these dependencies
explicitly.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>