Clarify which packages we must install in Rust Dockerfile, and why
`libssl-dev` and `pkg-config` are only required for compiling
`cargo-tarpaulin` (and nothing else), so we'll move those
installations to the `tarpaulin` target to localize them better.

`build-essential` and `zlib1g-dev` appear to be superfluous at this
point, so they have been removed.

Finally, the reason why we install the rest of the packages is shared
in a comment.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>