Reorganize "verify" Buildkite pipeline with the new "groups" feature
Our verify pipeline was getting a bit scattered and
disorganized. Buildkite also recently released a new "groups" feature
for pipelines to help with this exact scenario.

Thus, we now have groups :)

See https://buildkite.com/blog/announcing-the-new-group-step-type for
additional details.

(The only "bad" thing about this is with the audit jobs... due to how
the `chronotc/monorepo-diff` plugin currently works, group membership
cannot be propagated into the jobs that are ultimately uploaded to the
pipeline. This makes sense, though, because the groups functionality
is new. An issue has been filed, though:
https://github.com/chronotc/monorepo-diff-buildkite-plugin/issues/82. Despite
this, I still think that the organizational benefits gained are worth
it. As an alternative, we could just run the audit checks
unconditionally, or relegate them to a nightly-only pipeline.)

(I'm not touching the other pipelines because they don't appear to
benefit much from groups at this time.)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>