List targets to be ignored by `pants tailor` command
Running `./pants tailor` will automatically generate `BUILD` files
and/or targets for files not currently covered by Pants. This makes it
easy to ensure that new files added to the repository are properly
covered by Pants. However, not _everything_ really needs to be exposed
to Pants this way. In these cases, we can add corresponding targets to
the `ignore_adding_targets` configuration list in `pants.toml`.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>