Remove heredoc for modifying Nomad client `Cargo.toml`
As far as I can determine, we _must_ have an explicit `\n` in the
`sed` substitution we use to tweak the features that the generated
Nomad client uses. As such, there's really no benefit of using a
heredoc to define that replacement string; it just adds noise.

(Having a heredoc with an escaped newline in it is just an evil trap
for people like me that come along later to make it an actual
multiline string :sweat_smile:)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>