Internal 858 refactor rust proto (#1519)
These changes add Timestamp and Duration messages to our protocol buffers to replace google.protobuf.{Timestamp, Duration} because Google's types employ a signed duration, which leads to subtle complexities in time handling code. We sidestep these complexities by implementing our own unsigned types.

This PR also introduced a rust-proto-new crate, which is the beginning of a rewrite of the rust-proto crate. This rust-proto-new crate departs from rust-proto in that it encapsulates all the protobuf stubs, and presents a simple serialize/deserialize API to consumers. We'll gradually move all the rust code over to this new library as we migrate services to Kafka.