Allow passing in tracing endpoint with sane default
We allow passing in tracing endpoints so that this will work correctly
in AWS or when someone has multiple networks locally. That said, we
also default to what the private ip should be.
We're using IPs here instead of DNS, since consul dns inside containers
requires a fair amount of setup and management of computers. I suspect
that someone at grapl has already customized their dns setup so instead
we're using IP addresses as the default