Set service protocols explicitly for those with http
We're going to be setting the default protocol as grpc. As such, we
need to override that default for any service using http.
Also for services that were set to tcp *cough web-ui and ingress cough*
we're setting those to use http so we can get proper tracing and metrics