Put it all together in pulumi
This sets up the service configs, the proxy config for consul.
The service configs appear to be necessary to override the protocol
properly. It can be done in nomad, but it doesn't do everything
required (WTF nomad!)
The proxy config is set via pulumi. This way we have a fast way to
update the proxy config if necessary instead of having to redeploy consul
with a rolling deploy via autoscaling.
We do require the service configs to be created prior to the nomad job
otherwise you can end up in a bad state where the consul configs fail