Pass in eth0 IP from makefile to pulumi container
Since we want the private IP of the jaeger container, we can't get it
from within the pulumi container. Since the private IP should always be
the default IP address while we bind to 0.0.0.0 per the docs, we grab
the eth0 ip address on the assumption that should be the default IP.
We then pass that in to the pulumi container.