Remove LOCALSTACK_HOST from local-grapl.env, and associated code
This wasn't used directly anymore (I've removed the indirection in the
creation of `GRAPL_AWS_ENDPOINT`). We also had some other code that
referenced `LOCALSTACK_HOST`, but it was related to lambdas running in
Localstack. Since we no longer _have_ lambdas running in Localstack,
that code can be dropped as well.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>