Remove the --load parameter for docker build
The load parameter outputs a tarball. When using the docker-container
driver this is required for the built images to show up in `docker
image ls`, however it does slow down the builds. Since tracing docker
builds will be used primarily for debugging docker problems and build
performance, this seems like a useful optimization